requirements.txt
streamlit==1.29.0
pdfplumber==0.10.0
requests==2.31.0
beautifulsoup4==4.12.2
spacy==3.7.2
langchain==0.1.0
langchain-openai==0.0.2
python-dotenv==1.0.0
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.26.3

